Normalize help messages in examples
This makes `onlyargs_derive` output only a single blank line at the end
of the help message instead of two (which was characteristic of the
previous release). Now the output is closer between the manual trait
implementation and the derived trait impl.

Also changes the unordered list in the footer to use asterisks, which
looks a bit better in ASCII and disambiguates between flags and options.